src/_newtestrequirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas
boto3==1.18.21
botocore==1.21.41
brotlipy==0.7.0
bzip2==1.0.8
ca-certificates==2021.10.26
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
cryptography==35.0.0
ffmpeg==4.3
filelock==3.3.2
giflib==5.2.1
idna==3.2
importlib-metadata==4.8.1
jmespath==0.10.0
joblib==1.1.0
jpeg==9d
lcms2==2.12
libffi==3.3
libiconv==1.15
libpng==1.6.37
libtiff==4.2.0
libuv==1.40.0
libwebp==1.2.0
libwebp-base==1.2.0
lz4-c==1.9.3
mkl==2021.4.0
mkl-service==2.4.0
mkl_fft==1.3.1
mkl_random==1.2.2
ninja==1.10.2
numpy==1.21.2
numpy-base==1.21.2
olefile==0.46
openssl==1.1.1l
packaging==21.2
pillow==8.4.0
pip==21.0.1
pycparser==2.20
pyparsing==2.4.7
pysocks==1.7.1
python-dateutil==2.8.2
pytorch==1.7.0
#pytorch-mutex==1.0
pyyaml==6.0
quadprog==0.1.10
regex==2021.8.3
s3transfer==0.5.0
sacremoses==0.0.43
scikit-learn==1.0.1
scipy==1.7.1
threadpoolctl==2.2.0
tk==8.6.11
tokenizers==0.10.3
#torchaudio==0.7.0
#torchvision==0.8.1
transformers==4.10.2
typing_extensions==3.10.0.2
xz==5.2.5
zipp==3.6.0
zlib==1.2.11
zstd==1.4.9
pandas



